requirements.txt
discord.py[voice]
asyncio
yt_dlp
pathlib
